+++ lib/scanner/min_max_heap.c
@@ -15,6 +15,7 @@
 #include "../../utilities/dups_list.h"
 #include "../btree/btree.h"
 #include "../btree/conf.h"
+#include "../btree/key_splice.h"
 #include "../btree/kv_pairs.h"
 #include "../common/common.h"
 #include <assert.h>
@@ -56,7 +57,7 @@ static void push_back_duplicate_kv(struct sh_heap *heap, struct sh_heap_node *hp
 	uint64_t segment_offset =
 		ABSOLUTE_ADDRESS(keyvalue->dev_offt) - (ABSOLUTE_ADDRESS(keyvalue->dev_offt) % SEGMENT_SIZE);
 	struct kv_splice *kv = (struct kv_splice *)keyvalue->dev_offt;
-	assert(get_key_size(kv) <= MAX_KEY_SPLICE_SIZE);
+	assert(get_key_size(kv) <= (int32_t)MAX_KEY_SPLICE_SIZE);
 	struct dups_node *node = find_element(heap->dups, (uint64_t)REAL_ADDRESS(segment_offset));
 
 	if (node)
